[!!!][TASK] Lets TYPO3 create needed database fields automatically
Some minor deviations in the database field configuration might occur.
In most cases this should not be a problem.

This commit also contains a update wizard to migrate the related posts
field in the pages table.

Before this wizard would be executed, the database compare for this
field will fail.